Backport upstream patch to also match on 16 char TASK_COMM_LEN
At some point before the buster release, killall changed its COMM_LEN to
64 to cope with an (expected?) similar change in the kernel. But this
change is not present in buster's kernel and in fact, in none of the
other newer kernels AFAIK (likely due to the kernel "no-user-space
breakage" rule). And due to this mismatch, killall is not able to
identify process with names bigger than 15 characters.

Newer versions of killall thus learned to handle both possible lenghts
and this commit just backports the corresponding patch to fix this
regression in Debian 10.